add color to thermometer

+++ moralpain/lib/thermometer/cubit/thermometer_cubit.dart
@@ -3,9 +3,9 @@ import 'package:bloc/bloc.dart';
 /// {@template counter_cubit}
 /// A [Cubit] which manages an [int] as its state.
 /// {@endtemplate}
-class ThermometerCubit extends Cubit<int> {
+class ThermometerCubit extends Cubit<double> {
   /// {@macro counter_cubit}
   ThermometerCubit() : super(0);
 
-  void set(int val) => emit(val);
+  void set(double val) => emit(val);
 }
